+++ changelog.md
@@ -2,4 +2,5 @@
 ## 0.3.0
 
 - rename `files` field to `resources` in catalog files
+- all data files are stored under `resources` instead of `sources` and `data`
 - `id` subcommand is now `search`